requirements.txt
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
django-tastypie==0.13.3
django-wysiwyg-redactor==0.4.9
html-table-parser==0.1.0
